refactor: move service-worker-integration to config dir and make the build fail if building the service worker fails
