+++ astro.config.ts
@@ -5,7 +5,7 @@ import sitemap from '@astrojs/sitemap';
 import type { PluginOption } from 'vite';
 import { isPreview } from './config/preview';
 import pkg from './package.json';
-import serviceWorker from './scripts/service-worker-integration.ts';
+import serviceWorker from './config/astro/service-worker-integration.ts';
 
 const productionUrl = `https://${ pkg.name }.pages.dev`; // overwrite if you have a custom domain
 const localhostPort = 4323; // 4323 is "head" in T9
